changed doubles to floats

+++ LorenzAttractorPatch.hpp
@@ -9,11 +9,11 @@
  */
 class LorenzAttractorPatch : public Patch {
 public:
-  double x, y, z;
+  float x, y, z;
   float dt;
-  const double a = 10.0;
-  const double b = 8.0/3.0;
-  const double c = 28.0;
+  const float a = 10.0;
+  const float b = 8.0/3.0;
+  const float c = 28.0;
 
   LorenzAttractorPatch(){
     registerParameter(PARAMETER_A, "Rate");
@@ -32,12 +32,12 @@ public:
     if(isButtonPressed(PUSHBUTTON))
       reset();
     dt = getParameterValue(PARAMETER_A)*getParameterValue(PARAMETER_A)*0.0250;
-    double gainL = getParameterValue(PARAMETER_C)*2/25.0;
-    double gainR = getParameterValue(PARAMETER_D)*2/25.0;
+    float gainL = getParameterValue(PARAMETER_C)*2/25.0;
+    float gainR = getParameterValue(PARAMETER_D)*2/25.0;
     int size = buffer.getSize();
     float* left = buffer.getSamples(0);
     float* right = buffer.getSamples(1);
-    double dx, dy, dz;
+    float dx, dy, dz;
     for(int i=0;i<size;i++){
       dx = a*(y - x);
       dy = (x * (c - z) - y);
